requirements.txt
fastapi==0.115.0
hypercorn==0.17.3
pillow==11.0.0
